+++ mdsf/src/tools/snakefmt.rs
@@ -19,3 +19,44 @@ pub const COMMANDS: [CommandType; 3] = [
 ];
 
 pub const IS_STDIN: bool = false;
+
+#[cfg(test)]
+mod test_snakefmt {
+    #[test_with::executable(snakefmt || pipx || uv)]
+    fn test_snakefmt_snakemake_cdccd086422a6b0a() {
+        let input = r#"from snakemake.utils import min_version
+min_version("5.14.0")
+configfile: "config.yaml"
+include: "rules/foo.smk"
+"#;
+
+        let output = r#"from snakemake.utils import min_version
+
+min_version("5.14.0")
+
+
+configfile: "config.yaml"
+
+
+include: "rules/foo.smk"
+"#;
+
+        let file_ext = crate::fttype::get_file_extension("snakemake");
+
+        let snippet =
+            crate::execution::setup_snippet(input, &file_ext).expect("it to create a snippet file");
+
+        let result = crate::tools::Tooling::Snakefmt
+            .format_snippet(
+                snippet.path(),
+                crate::testing::DEFAULT_TEST_FORMATTER_TIMEOUT,
+                crate::testing::DEFAULT_TEST_DEBUG_ENABLED,
+                &crate::config::MdsfConfigRunners::all(),
+            )
+            .expect("it to be successful")
+            .1
+            .expect("it to be some");
+
+        assert_eq!(result, output);
+    }
+}