+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/hougesen/mdsf/compare/v0.8.5...HEAD)
 
+- test(tools): validate snakefmt [`#944`](https://github.com/hougesen/mdsf/pull/944)
+- test: disabling runners in config works [`#943`](https://github.com/hougesen/mdsf/pull/943)
 - test: format --config path argument [`#942`](https://github.com/hougesen/mdsf/pull/942)
 - refactor(cli): make --log-level a global argument [`#941`](https://github.com/hougesen/mdsf/pull/941)
 - feat(init): add --force argument [`#940`](https://github.com/hougesen/mdsf/pull/940)